converting features to float instead of strings

+++ pipeline/extractor/xbi_extractor.py
@@ -14,7 +14,7 @@ class XBIExtractor():
         for feature in self._features:
             Xt.append(arff_data['data'][:,attributes.index(feature)])
 
-        arff_data['X'] = np.array(Xt).T
+        arff_data['X'] = np.array(Xt, dtype='float64').T
         arff_data['y'] = np.array(arff_data['data'][:,attributes.index(self._class_label)])
 
         return arff_data
